Direct download for send (#243)
* Remove Get file capability

This needs to be removed because the SendFileResponse no longer contains
a url to download the file from. Instead, a GetDownloadLink method
must be used. That method increments access count, which is not
desirable for the owner of the Send. The cleanest approach is to remove
the capability, which also matches Web client's behavior

* jslib updates

* Use GetDownloadData method to receive download Url

* Update jslib