Add `bw get notes <id>` command (#255)
Rationale: the notes object is a freeform plain-text field that's
prominently displayed in Web Vault. It is also useful for the CLI users
as discussed before in issues #81 and #196. I have some use cases
planned myself.

I was rather surprised this wasn't supported already, but the
implementation is simple and cannot really break any existing
functionality so here it is.