[PoC] User Password Access Report
This proof of concept provides a report over the passwords a given user
has accessed during a specified period of time which address organization users.
The difference between the standard `Event Log` provided through the API or
the user interface (web) is that the `Event Log` provides a kind of audit log
where each interaction is logged and shown as it happens. This results in many cases
to multiple log entries showing the same information (user has view item xy) which
is already really helpful.

There is another scenario, where it would be of interest to a consolidated report of
total items a user has viewed which should be covered with the proof of concept
provided within this merge request.

The report shows each password item just once even the user has viewed it many times.